requirements.txt
mysql-connector-python~=9.3.0
selenium~=4.34.2
openpyxl~=3.1.5
pandas~=2.3.1
schedule~=1.2.2